+++ src/components/ContextProvider.tsx
@@ -13,7 +13,7 @@ import type { settings } from "../types/settings";
 import { api } from "../lib/api";
 import jwtDecode from "jwt-decode";
 import { AlertDialogProps } from "../lib/alertDialog";
-import { Category, User } from "@metahkg/api";
+import { BlockedUser, Category, User, Star } from "@metahkg/api";
 
 const Context = createContext<{
     back: [string, Dispatch<SetStateAction<string>>];
@@ -28,7 +28,8 @@ const Context = createContext<{
     user: [User | null, Dispatch<SetStateAction<User | null>>];
     alertDialog: [AlertDialogProps, Dispatch<SetStateAction<AlertDialogProps>>];
     reCaptchaSiteKey: string;
-    blockList: [User[], Dispatch<SetStateAction<User[]>>];
+    blockList: [BlockedUser[], Dispatch<SetStateAction<BlockedUser[]>>];
+    starList: [Star[], Dispatch<SetStateAction<Star[]>>];
     //@ts-ignore
 }>(null);
 /**
@@ -88,12 +89,27 @@ export default function ContextProvider(props: {
         message: "",
         btns: [],
     });
-    const [blockList, setBlockList] = useState<User[]>(
+    const [blockList, setBlockList] = useState<BlockedUser[]>(
         JSON.parse(localStorage.getItem("blocklist") || "[]")
+    );
+    const [starList, setStarList] = useState<Star[]>(
+        JSON.parse(localStorage.getItem("starlist") || "[]")
     );
 
     useEffect(() => {
         api.categories().then(setCategories);
+        if (user) {
+            api.meBlocked().then(setBlockList);
+            setInterval(() => {
+                api.meBlocked().then(setBlockList);
+            }, 1000 * 60 * 10);
+
+            api.meStarred().then(setStarList);
+            setInterval(() => {
+                api.meStarred().then(setStarList);
+            }, 1000 * 60 * 10);
+        }
+        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
     useEffect(() => {
@@ -138,6 +154,7 @@ export default function ContextProvider(props: {
                 reCaptchaSiteKey,
                 alertDialog: [alertDialog, setAlertDialog],
                 blockList: [blockList, setBlockList],
+                starList: [starList, setStarList],
             }}
         >
             {props.children}
@@ -261,4 +278,9 @@ export function useAlertDialog() {
 export function useBlockList() {
     const { blockList } = useContext(Context);
     return blockList;
+}
+
+export function useStarList() {
+    const { starList } = useContext(Context);
+    return starList;
 }