+++ src/components/sidebar.tsx
@@ -19,6 +19,7 @@ import {
     Menu as MenuIcon,
     Settings as SettingsIcon,
     Telegram as TelegramIcon,
+    Star as StarIcon,
 } from "@mui/icons-material";
 import { Link } from "react-router-dom";
 import { useNavigate } from "react-router-dom";
@@ -117,6 +118,11 @@ export default function SideBar() {
                                 link: "/recall",
                                 icon: <AccessTimeFilled />,
                             },
+                            user && {
+                                title: "Starred",
+                                link: "/starred",
+                                icon: <StarIcon />,
+                            },
                             {
                                 title: user ? "Logout" : "Login / Register",
                                 link: `/users/${
@@ -124,18 +130,21 @@ export default function SideBar() {
                                 }?returnto=${encodeURIComponent(wholePath())}`,
                                 icon: user ? <LogoutIcon /> : <AccountCircleIcon />,
                             },
-                        ].map((item, index) => (
-                            <ListItemButton
-                                key={index}
-                                component={Link}
-                                onClick={onClick}
-                                to={item.link}
-                                className="!no-underline text-white"
-                            >
-                                <ListItemIcon>{item.icon}</ListItemIcon>
-                                <ListItemText>{item.title}</ListItemText>
-                            </ListItemButton>
-                        ))}
+                        ].map(
+                            (item, index) =>
+                                item && (
+                                    <ListItemButton
+                                        key={index}
+                                        component={Link}
+                                        onClick={onClick}
+                                        to={item.link}
+                                        className="!no-underline text-white"
+                                    >
+                                        <ListItemIcon>{item.icon}</ListItemIcon>
+                                        <ListItemText>{item.title}</ListItemText>
+                                    </ListItemButton>
+                                )
+                        )}
                     </List>
                     <Divider />
                     {[